--- apps/etterlatte-behandling/src/main/kotlin/behandling/domain/BehandlingMedGrunnlagsopplysning.kt
@@ -1,12 +1,0 @@
-package no.nav.etterlatte.behandling.domain
-
-import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
-import no.nav.etterlatte.sak.SakMedUtlandstilknytning
-import java.time.LocalDateTime
-
-data class BehandlingMedGrunnlagsopplysning<Person>(
-    val behandling: Behandling,
-    val soeknadMottattDato: LocalDateTime?,
-    val personopplysning: Grunnlagsopplysning<Person>?,
-    val utenlandstilknytning: SakMedUtlandstilknytning?,
-)
