+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/utlandstilknytning/UtlandstilknytningVurdering.tsx
